refactor: add tests for service installations

+++ scripts/install.sh
@@ -3,11 +3,11 @@
 set -e
 set -x
 
-PROJECT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." >/dev/null 2>&1 && pwd )"
+PROJECT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." >/dev/null 2>&1 && pwd)"
 
-if [ "$EUID" -eq 0 ]
-  then echo "Please run as current user"
-  exit
+if [ "$EUID" -eq 0 ]; then
+    echo "Please run as current user"
+    exit
 fi
 
 WYOMING_SATELLITE=${WYOMING_SATELLITE:-1}
@@ -17,25 +17,32 @@ LIBRESPOT=${LIBRESPOT:-0}
 sudo apt install -y pulseaudio
 mkdir -p $HOME/src
 
+# Check if $HOME/src/wyoming-satellite directory doesn't exist
 # Install Wyoming satellite
 if [[ $WYOMING_SATELLITE == 1 ]]; then
-    git clone https://github.com/rhasspy/wyoming-satellite.git $HOME/src/.
-    cp $PROJECT_DIR/wyoming-satellite/sounds/* $HOME/src/wyoming-satellite/sounds/
+    if [ ! -d "$HOME/src/wyoming-satellite" ]; then
+        git clone https://github.com/rhasspy/wyoming-satellite.git $HOME/src/.
+        cp $PROJECT_DIR/wyoming-satellite/sounds/* $HOME/src/wyoming-satellite/sounds/
+    fi
 fi
 
 # Install Squeezelite manually (Debian package is dodgy)
 if [[ $SQUEEZELITE == 1 ]]; then
-    mkdir -p $HOME/src/squeezelite
-    cd $HOME/src/squeezelite
-    wget https://sourceforge.net/projects/lmsclients/files/squeezelite/linux/squeezelite-pulse-1.9.9.1428-aarch64.tar.gz/download
-    tar -xf download
-    cd $HOME
+    if [ ! -f "$HOME/src/squeezelite/squeezelite" ]; then
+        mkdir -p $HOME/src/squeezelite
+        cd $HOME/src/squeezelite
+        wget https://sourceforge.net/projects/lmsclients/files/squeezelite/linux/squeezelite-pulse-1.9.9.1428-aarch64.tar.gz/download
+        tar -xf download
+        cd $HOME
+    fi
 fi
 
 # librespot
 if [[ $LIBRESPOT == 1 ]]; then
-    sudo apt install -y build-essential cargo libpulse-dev
-    cargo install librespot --locked
+    if [ ! -f "$HOME/.cargo/bin/librespot" ]; then
+        sudo apt install -y build-essential cargo libpulse-dev
+        cargo install librespot --locked
+    fi
 fi
 
 # Install services
